+++ tap_salesforce/__init__.py
@@ -37,7 +37,6 @@ def main_impl():
             credentials=credentials,
             quota_percent_total=CONFIG.get("quota_percent_total"),
             quota_percent_per_run=CONFIG.get("quota_percent_per_run"),
-            is_sandbox=CONFIG.get("is_sandbox"),
             select_fields_by_default=CONFIG.get("select_fields_by_default"),
             default_start_date=CONFIG.get("start_date"),
             api_type=CONFIG.get("api_type"),